refactor(rust-converter): split ffi metadata and fuzz targets
