+++ components/rust-converter/src/converter.rs
@@ -104,9 +104,23 @@
 //! - Memory usage is proportional to output size, not input DOM size
 
 use crate::error::ConversionError;
+use html5ever::Attribute;
 use markup5ever_rcdom::{Handle, NodeData, RcDom};
 use std::cell::Ref;
 use std::time::{Duration, Instant};
+
+#[path = "converter/blocks.rs"]
+mod blocks;
+#[path = "converter/front_matter.rs"]
+mod front_matter;
+#[path = "converter/inline.rs"]
+mod inline;
+#[path = "converter/normalize.rs"]
+mod normalize;
+#[path = "converter/tables.rs"]
+mod tables;
+#[path = "converter/traversal.rs"]
+mod traversal;
 
 /// Markdown flavor selection
 #[derive(Debug, Clone, Copy)]
@@ -502,17 +516,7 @@ impl MarkdownConverter {
 
         // Extract metadata and add YAML front matter if enabled
         if self.options.include_front_matter && self.options.extract_metadata {
-            use crate::metadata::MetadataExtractor;
-
-            let extractor = MetadataExtractor::new(
-                self.options.base_url.clone(),
-                self.options.resolve_relative_urls,
-            );
-
-            if let Ok(metadata) = extractor.extract(dom) {
-                self.write_front_matter(&mut output, &metadata)?;
-            }
-
+            self.maybe_write_front_matter_from_dom(dom, &mut output)?;
             // Check timeout after metadata extraction
             ctx.check_timeout()?;
         }
@@ -531,1743 +535,6 @@ impl MarkdownConverter {
         ctx.check_timeout()?;
 
         Ok(markdown)
-    }
-
-    /// Write YAML front matter from metadata
-    ///
-    /// Generates a YAML front matter block with extracted metadata. The front matter
-    /// is enclosed in `---` delimiters and includes fields that have values.
-    ///
-    /// # YAML Formatting Rules
-    ///
-    /// - Only include fields that have non-empty values
-    /// - Escape YAML special characters in values (quotes, colons, etc.)
-    /// - Use double quotes for string values to ensure proper escaping
-    /// - Include resolved absolute URLs for images
-    ///
-    /// # Arguments
-    ///
-    /// * `output` - Mutable string buffer to write front matter to
-    /// * `metadata` - Extracted page metadata
-    ///
-    /// # Format
-    ///
-    /// ```yaml
-    /// ---
-    /// title: "Page Title"
-    /// url: "https://example.com/page"
-    /// description: "Page description"
-    /// image: "https://example.com/image.png"
-    /// author: "Author Name"
-    /// published: "2024-01-15"
-    /// ---
-    ///
-    /// ```
-    ///
-    /// # Requirements
-    ///
-    /// Validates: FR-15.3, FR-15.4, FR-15.5
-    fn write_front_matter(
-        &self,
-        output: &mut String,
-        metadata: &crate::metadata::PageMetadata,
-    ) -> Result<(), ConversionError> {
-        // Start YAML front matter block
-        output.push_str("---\n");
-
-        // Add title (required field per FR-15.4)
-        if let Some(ref title) = metadata.title
-            && !title.is_empty()
-        {
-            output.push_str("title: ");
-            self.write_yaml_string(output, title);
-            output.push('\n');
-        }
-
-        // Add URL (required field per FR-15.4)
-        if let Some(ref url) = metadata.url
-            && !url.is_empty()
-        {
-            output.push_str("url: ");
-            self.write_yaml_string(output, url);
-            output.push('\n');
-        }
-
-        // Add description (optional field per FR-15.5)
-        if let Some(ref description) = metadata.description
-            && !description.is_empty()
-        {
-            output.push_str("description: ");
-            self.write_yaml_string(output, description);
-            output.push('\n');
-        }
-
-        // Add image with resolved absolute URL (optional field per FR-15.5)
-        if let Some(ref image) = metadata.image
-            && !image.is_empty()
-        {
-            output.push_str("image: ");
-            self.write_yaml_string(output, image);
-            output.push('\n');
-        }
-
-        // Add author (optional field)
-        if let Some(ref author) = metadata.author
-            && !author.is_empty()
-        {
-            output.push_str("author: ");
-            self.write_yaml_string(output, author);
-            output.push('\n');
-        }
-
-        // Add published date (optional field)
-        if let Some(ref published) = metadata.published
-            && !published.is_empty()
-        {
-            output.push_str("published: ");
-            self.write_yaml_string(output, published);
-            output.push('\n');
-        }
-
-        // End YAML front matter block with blank line separator
-        output.push_str("---\n\n");
-
-        Ok(())
-    }
-
-    /// Write a YAML string value with proper escaping
-    ///
-    /// Escapes YAML special characters and wraps the value in double quotes.
-    /// This ensures the value is properly interpreted by YAML parsers.
-    ///
-    /// # YAML Special Characters
-    ///
-    /// The following characters require escaping:
-    /// - `"` (double quote) -> `\"`
-    /// - `\` (backslash) -> `\\`
-    /// - Newlines and control characters are preserved within quotes
-    ///
-    /// # Arguments
-    ///
-    /// * `output` - Mutable string buffer to write to
-    /// * `value` - String value to escape and write
-    ///
-    /// # Examples
-    ///
-    /// - `Hello World` -> `"Hello World"`
-    /// - `Title: Subtitle` -> `"Title: Subtitle"`
-    /// - `Quote "test"` -> `"Quote \"test\""`
-    fn write_yaml_string(&self, output: &mut String, value: &str) {
-        output.push('"');
-        for ch in value.chars() {
-            match ch {
-                '"' => output.push_str("\\\""),
-                '\\' => output.push_str("\\\\"),
-                '\n' => output.push_str("\\n"),
-                '\r' => output.push_str("\\r"),
-                '\t' => output.push_str("\\t"),
-                _ => output.push(ch),
-            }
-        }
-        output.push('"');
-    }
-
-    /// Returns true if the output buffer already contains Markdown body content.
-    ///
-    /// When YAML front matter is enabled, the output buffer is pre-populated before DOM
-    /// traversal starts. Text-node whitespace normalization should not treat that prefix
-    /// as body content, otherwise leading whitespace in the first body text node can be
-    /// emitted inconsistently depending on the front matter toggle.
-    fn has_body_content(&self, output: &str) -> bool {
-        if output.is_empty() {
-            return false;
-        }
-
-        if self.options.include_front_matter
-            && self.options.extract_metadata
-            && output.starts_with("---\n")
-            && let Some(rest) = output.strip_prefix("---\n")
-            && let Some(end_offset) = rest.find("\n---\n")
-        {
-            let body = &rest[end_offset + 5..];
-            return body.chars().any(|ch| !ch.is_whitespace());
-        }
-
-        true
-    }
-
-    /// Traverse a DOM node and convert it to Markdown
-    ///
-    /// This is the core recursive traversal function. It processes each node
-    /// according to its type and recursively processes children.
-    ///
-    /// # Arguments
-    ///
-    /// * `node` - Current DOM node to process
-    /// * `output` - Mutable string buffer for Markdown output
-    /// * `depth` - Current nesting depth (0 = document root)
-    ///
-    /// # Traversal Strategy
-    ///
-    /// The traversal follows these steps:
-    /// 1. Process the current node based on its type
-    /// 2. Recursively process all child nodes in document order
-    /// 3. Apply any closing formatting (e.g., blank lines after blocks)
-    ///
-    /// # Depth Tracking
-    ///
-    /// The depth parameter enables:
-    /// - Proper indentation for nested structures
-    /// - Detection of excessive nesting
-    /// - Context-aware formatting decisions
-    fn traverse_node(
-        &self,
-        node: &Handle,
-        output: &mut String,
-        depth: usize,
-    ) -> Result<(), ConversionError> {
-        match node.data {
-            NodeData::Document => {
-                // Document root - process all children
-                for child in node.children.borrow().iter() {
-                    self.traverse_node(child, output, depth)?;
-                }
-            }
-            NodeData::Element { ref name, .. } => {
-                // Element node - dispatch to appropriate handler
-                let tag_name = name.local.as_ref();
-                self.handle_element(node, tag_name, output, depth)?;
-            }
-            NodeData::Text { ref contents } => {
-                // Text node - extract and normalize text
-                let text = contents.borrow();
-                let normalized = self.normalize_text(&text);
-                if !normalized.is_empty() {
-                    // Add space before if original text had leading whitespace
-                    if text.starts_with(|c: char| c.is_whitespace())
-                        && self.has_body_content(output)
-                        && !output.ends_with(' ')
-                    {
-                        output.push(' ');
-                    }
-                    output.push_str(&normalized);
-                    // Add space after if original text had trailing whitespace
-                    if text.ends_with(|c: char| c.is_whitespace()) {
-                        output.push(' ');
-                    }
-                }
-            }
-            NodeData::Comment { .. } => {
-                // Comments are ignored in Markdown output
-            }
-            NodeData::Doctype { .. } => {
-                // DOCTYPE declarations are ignored
-            }
-            NodeData::ProcessingInstruction { .. } => {
-                // Processing instructions are ignored
-            }
-        }
-
-        Ok(())
-    }
-
-    /// Traverse a DOM node with timeout support
-    ///
-    /// This method is similar to `traverse_node` but includes cooperative timeout checking.
-    /// It increments the node count and checks timeout every 100 nodes.
-    ///
-    /// # Arguments
-    ///
-    /// * `node` - Current DOM node to process
-    /// * `output` - Mutable string buffer for Markdown output
-    /// * `depth` - Current nesting depth (0 = document root)
-    /// * `ctx` - Conversion context for timeout tracking
-    ///
-    /// # Timeout Checkpoints
-    ///
-    /// This method automatically checks timeout every 100 nodes by calling
-    /// `ctx.increment_and_check()`. This provides a balance between:
-    /// - Performance: Not checking on every single node
-    /// - Responsiveness: Detecting timeout within reasonable time
-    ///
-    /// # Requirements
-    ///
-    /// Validates: FR-10.2, FR-10.7
-    fn traverse_node_with_context(
-        &self,
-        node: &Handle,
-        output: &mut String,
-        depth: usize,
-        ctx: &mut ConversionContext,
-    ) -> Result<(), ConversionError> {
-        // Increment node count and check timeout at checkpoints (every 100 nodes)
-        ctx.increment_and_check()?;
-
-        match node.data {
-            NodeData::Document => {
-                // Document root - process all children
-                for child in node.children.borrow().iter() {
-                    self.traverse_node_with_context(child, output, depth, ctx)?;
-                }
-            }
-            NodeData::Element { ref name, .. } => {
-                // Element node - dispatch to appropriate handler
-                let tag_name = name.local.as_ref();
-                self.handle_element_with_context(node, tag_name, output, depth, ctx)?;
-            }
-            NodeData::Text { ref contents } => {
-                // Text node - extract and normalize text
-                let text = contents.borrow();
-                let normalized = self.normalize_text(&text);
-                if !normalized.is_empty() {
-                    // Add space before if original text had leading whitespace
-                    if text.starts_with(|c: char| c.is_whitespace())
-                        && self.has_body_content(output)
-                        && !output.ends_with(' ')
-                    {
-                        output.push(' ');
-                    }
-                    output.push_str(&normalized);
-                    // Add space after if original text had trailing whitespace
-                    if text.ends_with(|c: char| c.is_whitespace()) {
-                        output.push(' ');
-                    }
-                }
-            }
-            NodeData::Comment { .. } => {
-                // Comments are ignored in Markdown output
-            }
-            NodeData::Doctype { .. } => {
-                // DOCTYPE declarations are ignored
-            }
-            NodeData::ProcessingInstruction { .. } => {
-                // Processing instructions are ignored
-            }
-        }
-
-        Ok(())
-    }
-
-    /// Handle an HTML element and convert it to Markdown
-    ///
-    /// This function dispatches to specific element handlers based on the tag name.
-    /// It implements the element-specific conversion logic for each supported HTML element.
-    ///
-    /// # Arguments
-    ///
-    /// * `node` - The element node to process
-    /// * `tag_name` - The HTML tag name (e.g., "h1", "p", "div")
-    /// * `output` - Mutable string buffer for Markdown output
-    /// * `depth` - Current nesting depth
-    ///
-    /// # Supported Elements
-    ///
-    /// Currently supported elements:
-    /// - `h1` to `h6`: Headings (ATX-style)
-    /// - `p`: Paragraphs
-    /// - Other elements: Processed as containers (children traversed)
-    ///
-    /// # Future Extensions
-    ///
-    /// Additional element handlers will be added in subsequent tasks:
-    /// - Links (`a`)
-    /// - Images (`img`)
-    /// - Lists (`ul`, `ol`, `li`)
-    /// - Code blocks (`pre`, `code`)
-    /// - Formatting (`strong`, `em`, `code`)
-    /// - Tables (`table`, `tr`, `td`, `th`)
-    fn handle_element(
-        &self,
-        node: &Handle,
-        tag_name: &str,
-        output: &mut String,
-        depth: usize,
-    ) -> Result<(), ConversionError> {
-        // Security validation: check if element should be sanitized
-        use crate::security::SanitizeAction;
-        let sanitize_action = self.security_validator.check_element(tag_name);
-
-        match sanitize_action {
-            SanitizeAction::Remove => {
-                // Skip dangerous elements and their children
-                return Ok(());
-            }
-            SanitizeAction::Allow | SanitizeAction::StripAttributes | SanitizeAction::StripUrl => {
-                // Continue processing, but check attributes if needed
-            }
-        }
-
-        // Validate nesting depth
-        self.security_validator
-            .validate_depth(depth)
-            .map_err(ConversionError::InvalidInput)?;
-
-        match tag_name {
-            // Heading elements (h1-h6)
-            "h1" => self.handle_heading(node, 1, output, depth)?,
-            "h2" => self.handle_heading(node, 2, output, depth)?,
-            "h3" => self.handle_heading(node, 3, output, depth)?,
-            "h4" => self.handle_heading(node, 4, output, depth)?,
-            "h5" => self.handle_heading(node, 5, output, depth)?,
-            "h6" => self.handle_heading(node, 6, output, depth)?,
-
-            // Paragraph element
-            "p" => self.handle_paragraph(node, output, depth)?,
-
-            // Link element
-            "a" => self.handle_link(node, output, depth)?,
-
-            // Image element
-            "img" => self.handle_image(node, output, depth)?,
-
-            // List elements
-            "ul" => self.handle_list(node, output, 0, false)?,
-            "ol" => self.handle_list(node, output, 0, true)?,
-            "li" => self.handle_list_item(node, output, 0)?,
-
-            // Code elements
-            "pre" => self.handle_code_block(node, output, depth)?,
-            "code" => self.handle_inline_code(node, output, depth)?,
-
-            // Text formatting elements
-            "strong" | "b" => self.handle_bold(node, output, depth)?,
-            "em" | "i" => self.handle_italic(node, output, depth)?,
-
-            // Table elements (GFM only)
-            "table" => self.handle_table(node, output, depth)?,
-
-            // Elements to skip (non-content) - already handled by security validator
-            "script" | "style" | "noscript" => {
-                // Skip these elements and their children
-            }
-
-            // Default: process as container (traverse children)
-            _ => {
-                for child in node.children.borrow().iter() {
-                    self.traverse_node(child, output, depth + 1)?;
-                }
-            }
-        }
-
-        Ok(())
-    }
-
-    /// Handle an HTML element with timeout support
-    ///
-    /// This method is similar to `handle_element` but passes the conversion context
-    /// through to child traversals for timeout checking.
-    ///
-    /// # Arguments
-    ///
-    /// * `node` - The element node to process
-    /// * `tag_name` - The HTML tag name (e.g., "h1", "p", "div")
-    /// * `output` - Mutable string buffer for Markdown output
-    /// * `depth` - Current nesting depth
-    /// * `ctx` - Conversion context for timeout tracking
-    ///
-    /// # Requirements
-    ///
-    /// Validates: FR-10.2, FR-10.7
-    fn handle_element_with_context(
-        &self,
-        node: &Handle,
-        tag_name: &str,
-        output: &mut String,
-        depth: usize,
-        ctx: &mut ConversionContext,
-    ) -> Result<(), ConversionError> {
-        // Security validation: check if element should be sanitized
-        use crate::security::SanitizeAction;
-        let sanitize_action = self.security_validator.check_element(tag_name);
-
-        match sanitize_action {
-            SanitizeAction::Remove => {
-                // Skip dangerous elements and their children
-                return Ok(());
-            }
-            SanitizeAction::Allow | SanitizeAction::StripAttributes | SanitizeAction::StripUrl => {
-                // Continue processing, but check attributes if needed
-            }
-        }
-
-        // Validate nesting depth
-        self.security_validator
-            .validate_depth(depth)
-            .map_err(ConversionError::InvalidInput)?;
-
-        match tag_name {
-            // Heading elements (h1-h6)
-            "h1" => self.handle_heading(node, 1, output, depth)?,
-            "h2" => self.handle_heading(node, 2, output, depth)?,
-            "h3" => self.handle_heading(node, 3, output, depth)?,
-            "h4" => self.handle_heading(node, 4, output, depth)?,
-            "h5" => self.handle_heading(node, 5, output, depth)?,
-            "h6" => self.handle_heading(node, 6, output, depth)?,
-
-            // Paragraph element
-            "p" => self.handle_paragraph(node, output, depth)?,
-
-            // Link element
-            "a" => self.handle_link(node, output, depth)?,
-
-            // Image element
-            "img" => self.handle_image(node, output, depth)?,
-
-            // List elements
-            "ul" => self.handle_list(node, output, 0, false)?,
-            "ol" => self.handle_list(node, output, 0, true)?,
-            "li" => self.handle_list_item(node, output, 0)?,
-
-            // Code elements
-            "pre" => self.handle_code_block(node, output, depth)?,
-            "code" => self.handle_inline_code(node, output, depth)?,
-
-            // Text formatting elements
-            "strong" | "b" => self.handle_bold(node, output, depth)?,
-            "em" | "i" => self.handle_italic(node, output, depth)?,
-
-            // Table elements (GFM only)
-            "table" => self.handle_table(node, output, depth)?,
-
-            // Elements to skip (non-content) - already handled by security validator
-            "script" | "style" | "noscript" => {
-                // Skip these elements and their children
-            }
-
-            // Default: process as container (traverse children with context)
-            _ => {
-                for child in node.children.borrow().iter() {
-                    self.traverse_node_with_context(child, output, depth + 1, ctx)?;
-                }
-            }
-        }
-
-        Ok(())
-    }
-
-    /// Handle heading elements (h1-h6)
-    ///
-    /// Converts HTML headings to ATX-style Markdown headings using `#` symbols.
-    ///
-    /// # Arguments
-    ///
-    /// * `node` - The heading element node
-    /// * `level` - Heading level (1-6)
-    /// * `output` - Mutable string buffer for Markdown output
-    /// * `depth` - Current nesting depth
-    ///
-    /// # Output Format
-    ///
-    /// - Level 1: `# Heading`
-    /// - Level 2: `## Heading`
-    /// - Level 3: `### Heading`
-    /// - etc.
-    ///
-    /// Headings are followed by two newlines to create a blank line separator.
-    fn handle_heading(
-        &self,
-        node: &Handle,
-        level: usize,
-        output: &mut String,
-        depth: usize,
-    ) -> Result<(), ConversionError> {
-        // Ensure blank line before heading (if not at start)
-        if !output.is_empty() && !output.ends_with("\n\n") {
-            if output.ends_with('\n') {
-                output.push('\n');
-            } else {
-                output.push_str("\n\n");
-            }
-        }
-
-        // Add ATX-style heading markers
-        for _ in 0..level {
-            output.push('#');
-        }
-        output.push(' ');
-
-        // Process heading content (including inline elements like code)
-        let start_len = output.len();
-        for child in node.children.borrow().iter() {
-            self.traverse_node(child, output, depth + 1)?;
-        }
-
-        // Normalize the heading text (collapse whitespace, trim)
-        let heading_content = output[start_len..].to_string();
-        let normalized = self.normalize_text(&heading_content);
-        output.truncate(start_len);
-        output.push_str(&normalized);
-
-        // Add blank line after heading
-        output.push_str("\n\n");
-
-        Ok(())
-    }
-
-    /// Handle paragraph elements
-    ///
-    /// Converts HTML paragraphs to plain text with blank line separation.
-    ///
-    /// # Arguments
-    ///
-    /// * `node` - The paragraph element node
-    /// * `output` - Mutable string buffer for Markdown output
-    /// * `depth` - Current nesting depth
-    ///
-    /// # Output Format
-    ///
-    /// Paragraphs are rendered as plain text followed by two newlines to create
-    /// a blank line separator between paragraphs.
-    fn handle_paragraph(
-        &self,
-        node: &Handle,
-        output: &mut String,
-        depth: usize,
-    ) -> Result<(), ConversionError> {
-        // Ensure blank line before paragraph (if not at start)
-        if !output.is_empty() && !output.ends_with("\n\n") {
-            if output.ends_with('\n') {
-                output.push('\n');
-            } else {
-                output.push_str("\n\n");
-            }
-        }
-
-        // Process paragraph children (which may include inline elements like links, images)
-        let start_len = output.len();
-        for child in node.children.borrow().iter() {
-            self.traverse_node(child, output, depth + 1)?;
-        }
-
-        // Add blank line after paragraph if content was added
-        if output.len() > start_len {
-            output.push_str("\n\n");
-        }
-
-        Ok(())
-    }
-
-    /// Handle anchor (link) elements
-    ///
-    /// Converts HTML anchor tags to Markdown link format: `[text](url)`
-    ///
-    /// # Arguments
-    ///
-    /// * `node` - The anchor element node
-    /// * `output` - Mutable string buffer for Markdown output
-    /// * `depth` - Current nesting depth
-    ///
-    /// # Output Format
-    ///
-    /// Links are rendered as `[link text](href)` where:
-    /// - `link text` is the text content of the anchor element
-    /// - `href` is the value of the href attribute
-    ///
-    /// If the href attribute is missing, the link text is rendered as plain text.
-    ///
-    /// # Examples
-    ///
-    /// ```html
-    /// <a href="https://example.com">Example</a>
-    /// ```
-    /// becomes:
-    /// ```markdown
-    /// [Example](https://example.com)
-    /// ```
-    fn handle_link(
-        &self,
-        node: &Handle,
-        output: &mut String,
-        _depth: usize,
-    ) -> Result<(), ConversionError> {
-        // Extract href attribute
-        // Note: Attributes are processed in the order they appear in the DOM.
-        // For deterministic output, we rely on html5ever's consistent attribute ordering.
-        let href = if let NodeData::Element { ref attrs, .. } = node.data {
-            attrs
-                .borrow()
-                .iter()
-                .find(|attr| attr.name.local.as_ref() == "href")
-                .map(|attr| attr.value.to_string())
-        } else {
-            None
-        };
-
-        // Extract link text from children
-        let mut link_text = String::new();
-        for child in node.children.borrow().iter() {
-            self.extract_text(child, &mut link_text)?;
-        }
-        let normalized_text = self.normalize_text(&link_text);
-
-        // Generate Markdown link or plain text if no href
-        if let Some(url) = href {
-            // Security: Sanitize URL to prevent javascript: and data: URLs
-            if let Some(safe_url) = self.security_validator.sanitize_url(&url) {
-                if !normalized_text.is_empty() {
-                    output.push('[');
-                    output.push_str(&normalized_text);
-                    output.push_str("](");
-                    output.push_str(safe_url);
-                    output.push(')');
-                }
-            } else {
-                // Dangerous URL detected, render as plain text without link
-                if !normalized_text.is_empty() {
-                    output.push_str(&normalized_text);
-                }
-            }
-        } else {
-            // No href attribute, render as plain text
-            if !normalized_text.is_empty() {
-                output.push_str(&normalized_text);
-            }
-        }
-
-        Ok(())
-    }
-
-    /// Handle image elements
-    ///
-    /// Converts HTML img tags to Markdown image format: `![alt](src)`
-    ///
-    /// # Arguments
-    ///
-    /// * `node` - The img element node
-    /// * `output` - Mutable string buffer for Markdown output
-    /// * `depth` - Current nesting depth
-    ///
-    /// # Output Format
-    ///
-    /// Images are rendered as `![alt text](src)` where:
-    /// - `alt text` is the value of the alt attribute (or empty if missing)
-    /// - `src` is the value of the src attribute
-    ///
-    /// If the src attribute is missing, the image is not rendered.
-    ///
-    /// # Deterministic Output
-    ///
-    /// Attributes are processed in a consistent order (html5ever maintains insertion order)
-    /// to ensure deterministic output for stable ETag generation.
-    ///
-    /// # Examples
-    ///
-    /// ```html
-    /// <img src="image.png" alt="Description">
-    /// ```
-    /// becomes:
-    /// ```markdown
-    /// ![Description](image.png)
-    /// ```
-    fn handle_image(
-        &self,
-        node: &Handle,
-        output: &mut String,
-        _depth: usize,
-    ) -> Result<(), ConversionError> {
-        // Extract src and alt attributes
-        // Note: Attributes are processed in the order they appear in the DOM.
-        // For deterministic output, we rely on html5ever's consistent attribute ordering.
-        let (src, alt) = if let NodeData::Element { ref attrs, .. } = node.data {
-            let attrs_borrowed = attrs.borrow();
-            let src = attrs_borrowed
-                .iter()
-                .find(|attr| attr.name.local.as_ref() == "src")
-                .map(|attr| attr.value.to_string());
-            let alt = attrs_borrowed
-                .iter()
-                .find(|attr| attr.name.local.as_ref() == "alt")
-                .map(|attr| attr.value.to_string())
-                .unwrap_or_default();
-            (src, alt)
-        } else {
-            (None, String::new())
-        };
-
-        // Generate Markdown image if src is present and safe
-        if let Some(url) = src {
-            // Security: Sanitize URL to prevent javascript: and data: URLs
-            if let Some(safe_url) = self.security_validator.sanitize_url(&url) {
-                output.push_str("![");
-                output.push_str(&alt);
-                output.push_str("](");
-                output.push_str(safe_url);
-                output.push(')');
-            }
-            // If URL is dangerous, skip the image entirely
-        }
-
-        Ok(())
-    }
-
-    /// Handle list elements (ul/ol)
-    ///
-    /// Converts HTML unordered and ordered lists to Markdown list format.
-    ///
-    /// # Arguments
-    ///
-    /// * `node` - The list element node (ul or ol)
-    /// * `output` - Mutable string buffer for Markdown output
-    /// * `depth` - Current nesting depth
-    /// * `ordered` - true for ordered lists (ol), false for unordered lists (ul)
-    ///
-    /// # Output Format
-    ///
-    /// - Unordered lists use `- ` prefix
-    /// - Ordered lists use `1. ` prefix (all items numbered as 1)
-    /// - Nested lists are indented with 2 spaces per level
-    ///
-    /// # Examples
-    ///
-    /// ```html
-    /// <ul>
-    ///   <li>Item 1</li>
-    ///   <li>Item 2</li>
-    /// </ul>
-    /// ```
-    /// becomes:
-    /// ```markdown
-    /// - Item 1
-    /// - Item 2
-    /// ```
-    fn handle_list(
-        &self,
-        node: &Handle,
-        output: &mut String,
-        depth: usize,
-        ordered: bool,
-    ) -> Result<(), ConversionError> {
-        // Ensure blank line before list (if not at start)
-        if !output.is_empty() && !output.ends_with("\n\n") {
-            if output.ends_with('\n') {
-                output.push('\n');
-            } else {
-                output.push_str("\n\n");
-            }
-        }
-
-        // Store the list type in the context for list items
-        // Process all list item children
-        for child in node.children.borrow().iter() {
-            if let NodeData::Element { ref name, .. } = child.data
-                && name.local.as_ref() == "li"
-            {
-                self.handle_list_item_with_marker(child, output, depth, ordered)?;
-            }
-        }
-
-        // Ensure blank line after list
-        if !output.ends_with("\n\n") {
-            output.push('\n');
-        }
-
-        Ok(())
-    }
-
-    /// Handle list item elements (li)
-    ///
-    /// This is called when a list item is encountered outside of list context.
-    /// It delegates to handle_list_item_with_marker with default settings.
-    fn handle_list_item(
-        &self,
-        node: &Handle,
-        output: &mut String,
-        depth: usize,
-    ) -> Result<(), ConversionError> {
-        // Default to unordered list marker
-        self.handle_list_item_with_marker(node, output, depth, false)
-    }
-
-    /// Handle list item elements with specific marker type
-    ///
-    /// Converts HTML list items to Markdown list items with proper indentation.
-    ///
-    /// # Arguments
-    ///
-    /// * `node` - The list item element node
-    /// * `output` - Mutable string buffer for Markdown output
-    /// * `depth` - Current nesting depth (for indentation)
-    /// * `ordered` - true for ordered list marker (1.), false for unordered (-)
-    ///
-    /// # Output Format
-    ///
-    /// List items are indented based on depth:
-    /// - Depth 0: no indentation
-    /// - Depth 1: 2 spaces
-    /// - Depth 2: 4 spaces
-    /// - etc.
-    fn handle_list_item_with_marker(
-        &self,
-        node: &Handle,
-        output: &mut String,
-        depth: usize,
-        ordered: bool,
-    ) -> Result<(), ConversionError> {
-        // Add indentation based on depth (2 spaces per level)
-        for _ in 0..depth {
-            output.push_str("  ");
-        }
-
-        // Add list marker
-        if ordered {
-            output.push_str("1. ");
-        } else {
-            output.push_str("- ");
-        }
-
-        // Process list item content
-        let start_len = output.len();
-        for child in node.children.borrow().iter() {
-            match child.data {
-                NodeData::Element { ref name, .. } => {
-                    let tag_name = name.local.as_ref();
-                    // Handle nested lists
-                    if tag_name == "ul" {
-                        // Finish current line before nested list
-                        if output.len() > start_len && !output.ends_with('\n') {
-                            output.push('\n');
-                        }
-
-                        // Process nested unordered list
-                        self.handle_list(child, output, depth + 1, false)?;
-                    } else if tag_name == "ol" {
-                        // Finish current line before nested list
-                        if output.len() > start_len && !output.ends_with('\n') {
-                            output.push('\n');
-                        }
-
-                        // Process nested ordered list
-                        self.handle_list(child, output, depth + 1, true)?;
-                    } else {
-                        // Process other elements (including inline elements like <a>, <img>)
-                        self.traverse_node(child, output, depth + 1)?;
-                    }
-                }
-                _ => {
-                    // Process text nodes and other content
-                    self.traverse_node(child, output, depth + 1)?;
-                }
-            }
-        }
-
-        // Ensure line ends with newline
-        if !output.ends_with('\n') {
-            output.push('\n');
-        }
-
-        Ok(())
-    }
-
-    /// Handle code block elements (pre/code)
-    ///
-    /// Converts HTML code blocks to fenced code blocks in Markdown.
-    /// Detects language from class attributes (e.g., class="language-python").
-    /// Preserves code content without any text normalization.
-    ///
-    /// # Arguments
-    ///
-    /// * `node` - The pre element node
-    /// * `output` - Mutable string buffer for Markdown output
-    /// * `depth` - Current nesting depth
-    ///
-    /// # Output Format
-    ///
-    /// - With language: ```python\ncode\n```
-    /// - Without language: ```\ncode\n```
-    ///
-    /// Code blocks are surrounded by blank lines for proper separation.
-    fn handle_code_block(
-        &self,
-        node: &Handle,
-        output: &mut String,
-        _depth: usize,
-    ) -> Result<(), ConversionError> {
-        // Ensure blank line before code block (if not at start)
-        if !output.is_empty() && !output.ends_with("\n\n") {
-            if output.ends_with('\n') {
-                output.push('\n');
-            } else {
-                output.push_str("\n\n");
-            }
-        }
-
-        // Try to detect language from class attribute
-        let mut language = String::new();
-
-        // Check if this pre contains a code element with language class
-        for child in node.children.borrow().iter() {
-            if let NodeData::Element {
-                ref name,
-                ref attrs,
-                ..
-            } = child.data
-                && name.local.as_ref() == "code"
-            {
-                // Look for class attribute with language- prefix
-                for attr in attrs.borrow().iter() {
-                    if attr.name.local.as_ref() == "class" {
-                        let class_value = attr.value.to_string();
-                        // Look for language-* or lang-* patterns
-                        for class in class_value.split_whitespace() {
-                            if let Some(lang) = class.strip_prefix("language-") {
-                                language = lang.to_string();
-                                break;
-                            } else if let Some(lang) = class.strip_prefix("lang-") {
-                                language = lang.to_string();
-                                break;
-                            }
-                        }
-                        if !language.is_empty() {
-                            break;
-                        }
-                    }
-                }
-            }
-        }
-
-        // Start fenced code block
-        output.push_str("```");
-        if !language.is_empty() {
-            output.push_str(&language);
-        }
-        output.push('\n');
-
-        // Extract code content WITHOUT normalization
-        // This is critical - code must be preserved exactly as-is
-        self.extract_code_content(node, output)?;
-
-        // End fenced code block
-        // Ensure code ends with newline before closing fence
-        if !output.ends_with('\n') {
-            output.push('\n');
-        }
-        output.push_str("```");
-        output.push('\n');
-
-        // Ensure blank line after code block
-        output.push('\n');
-
-        Ok(())
-    }
-
-    /// Handle inline code elements (code)
-    ///
-    /// Converts HTML inline code to backtick-wrapped code in Markdown.
-    /// Preserves code content without modification.
-    ///
-    /// # Arguments
-    ///
-    /// * `node` - The code element node
-    /// * `output` - Mutable string buffer for Markdown output
-    /// * `depth` - Current nesting depth
-    ///
-    /// # Output Format
-    ///
-    /// - Inline code: `code`
-    ///
-    /// # Note
-    ///
-    /// This handler is only called for standalone code elements (inline code).
-    /// Code elements inside pre elements are handled by handle_code_block.
-    fn handle_inline_code(
-        &self,
-        node: &Handle,
-        output: &mut String,
-        _depth: usize,
-    ) -> Result<(), ConversionError> {
-        // Extract code content WITHOUT normalization
-        let mut code_content = String::new();
-        self.extract_code_content(node, &mut code_content)?;
-
-        // Wrap in backticks
-        output.push('`');
-        output.push_str(&code_content);
-        output.push('`');
-
-        Ok(())
-    }
-
-    /// Handle bold/strong elements
-    ///
-    /// Converts HTML bold elements (strong, b) to Markdown bold format: `**text**`
-    ///
-    /// # Arguments
-    ///
-    /// * `node` - The bold element node
-    /// * `output` - Mutable string buffer for Markdown output
-    /// * `depth` - Current nesting depth
-    ///
-    /// # Output Format
-    ///
-    /// Bold text is rendered as `**text**` where text is the content of the element.
-    /// Nested formatting is supported (e.g., bold within italic or vice versa).
-    ///
-    /// # Examples
-    ///
-    /// ```html
-    /// <strong>bold text</strong>
-    /// <b>also bold</b>
-    /// ```
-    /// becomes:
-    /// ```markdown
-    /// **bold text**
-    /// **also bold**
-    /// ```
-    fn handle_bold(
-        &self,
-        node: &Handle,
-        output: &mut String,
-        depth: usize,
-    ) -> Result<(), ConversionError> {
-        // Add opening bold marker
-        output.push_str("**");
-
-        // Process children (which may include nested formatting)
-        for child in node.children.borrow().iter() {
-            self.traverse_node(child, output, depth + 1)?;
-        }
-
-        // Add closing bold marker
-        output.push_str("**");
-
-        Ok(())
-    }
-
-    /// Handle italic/emphasis elements
-    ///
-    /// Converts HTML italic elements (em, i) to Markdown italic format: `*text*`
-    ///
-    /// # Arguments
-    ///
-    /// * `node` - The italic element node
-    /// * `output` - Mutable string buffer for Markdown output
-    /// * `depth` - Current nesting depth
-    ///
-    /// # Output Format
-    ///
-    /// Italic text is rendered as `*text*` where text is the content of the element.
-    /// Nested formatting is supported (e.g., italic within bold or vice versa).
-    ///
-    /// # Examples
-    ///
-    /// ```html
-    /// <em>italic text</em>
-    /// <i>also italic</i>
-    /// ```
-    /// becomes:
-    /// ```markdown
-    /// *italic text*
-    /// *also italic*
-    /// ```
-    fn handle_italic(
-        &self,
-        node: &Handle,
-        output: &mut String,
-        depth: usize,
-    ) -> Result<(), ConversionError> {
-        // Add opening italic marker
-        output.push('*');
-
-        // Process children (which may include nested formatting)
-        for child in node.children.borrow().iter() {
-            self.traverse_node(child, output, depth + 1)?;
-        }
-
-        // Add closing italic marker
-        output.push('*');
-
-        Ok(())
-    }
-
-    /// Handle table elements (GFM only)
-    ///
-    /// Converts HTML tables to GitHub Flavored Markdown table format.
-    /// Only enabled when GFM flavor is configured.
-    ///
-    /// # Arguments
-    ///
-    /// * `node` - The table element node
-    /// * `output` - Mutable string buffer for Markdown output
-    /// * `depth` - Current nesting depth
-    ///
-    /// # Output Format
-    ///
-    /// GFM tables use pipe separators:
-    /// ```markdown
-    /// | Header 1 | Header 2 |
-    /// | -------- | -------- |
-    /// | Cell 1   | Cell 2   |
-    /// ```
-    ///
-    /// Alignment is detected from style/align attributes:
-    /// - Left: `| :--- |` (default)
-    /// - Center: `| :---: |`
-    /// - Right: `| ---: |`
-    ///
-    /// # GFM Flavor Check
-    ///
-    /// Tables are only converted when flavor is GitHubFlavoredMarkdown.
-    /// For CommonMark, tables are processed as regular containers.
-    fn handle_table(
-        &self,
-        node: &Handle,
-        output: &mut String,
-        depth: usize,
-    ) -> Result<(), ConversionError> {
-        // Only convert tables for GFM flavor
-        if !matches!(self.options.flavor, MarkdownFlavor::GitHubFlavoredMarkdown) {
-            // For CommonMark, process as container (traverse children)
-            for child in node.children.borrow().iter() {
-                self.traverse_node(child, output, depth + 1)?;
-            }
-            return Ok(());
-        }
-
-        // Ensure blank line before table
-        if !output.is_empty() && !output.ends_with("\n\n") {
-            if output.ends_with('\n') {
-                output.push('\n');
-            } else {
-                output.push_str("\n\n");
-            }
-        }
-
-        // Extract table structure
-        let mut headers: Vec<String> = Vec::new();
-        let mut alignments: Vec<TableAlignment> = Vec::new();
-        let mut rows: Vec<Vec<String>> = Vec::new();
-
-        // Parse table children (thead, tbody, tr)
-        for child in node.children.borrow().iter() {
-            if let NodeData::Element { ref name, .. } = child.data {
-                match name.local.as_ref() {
-                    "thead" => {
-                        self.extract_table_header(child, &mut headers, &mut alignments)?;
-                    }
-                    "tbody" => {
-                        // Check if first row in tbody should be treated as header
-                        // If no headers yet, check if tbody's first row should be treated as header
-                        if headers.is_empty() {
-                            // Look for first tr in tbody
-                            let children = child.children.borrow();
-                            let first_tr_opt = children.iter().find(|c| {
-                                if let NodeData::Element { ref name, .. } = c.data {
-                                    name.local.as_ref() == "tr"
-                                } else {
-                                    false
-                                }
-                            });
-
-                            if let Some(first_tr) = first_tr_opt {
-                                // Check if first row has th elements
-                                let has_th = first_tr.children.borrow().iter().any(|c| {
-                                    if let NodeData::Element { ref name, .. } = c.data {
-                                        name.local.as_ref() == "th"
-                                    } else {
-                                        false
-                                    }
-                                });
-
-                                // Treat first row as header if it has th elements OR if it's the only way to get headers
-                                // (This handles cases where HTML uses td for headers)
-                                if has_th {
-                                    // First row is header (has th elements)
-                                    self.extract_table_row_as_header(
-                                        first_tr,
-                                        &mut headers,
-                                        &mut alignments,
-                                    )?;
-                                    // Extract remaining rows as data
-                                    let mut is_first = true;
-                                    for tbody_child in children.iter() {
-                                        if let NodeData::Element { ref name, .. } = tbody_child.data
-                                            && name.local.as_ref() == "tr"
-                                        {
-                                            if is_first {
-                                                is_first = false;
-                                                continue; // Skip header row
-                                            }
-                                            let mut row_cells = Vec::new();
-                                            self.extract_table_row(tbody_child, &mut row_cells)?;
-                                            rows.push(row_cells);
-                                        }
-                                    }
-                                } else {
-                                    // First row uses td but treat as header anyway (common pattern)
-                                    self.extract_table_row_as_header(
-                                        first_tr,
-                                        &mut headers,
-                                        &mut alignments,
-                                    )?;
-                                    // Extract remaining rows as data
-                                    let mut is_first = true;
-                                    for tbody_child in children.iter() {
-                                        if let NodeData::Element { ref name, .. } = tbody_child.data
-                                            && name.local.as_ref() == "tr"
-                                        {
-                                            if is_first {
-                                                is_first = false;
-                                                continue; // Skip header row
-                                            }
-                                            let mut row_cells = Vec::new();
-                                            self.extract_table_row(tbody_child, &mut row_cells)?;
-                                            rows.push(row_cells);
-                                        }
-                                    }
-                                }
-                            } else {
-                                // No rows in tbody
-                                self.extract_table_rows(child, &mut rows)?;
-                            }
-                        } else {
-                            // Headers already extracted from thead, all tbody rows are data
-                            self.extract_table_rows(child, &mut rows)?;
-                        }
-                    }
-                    "tr" => {
-                        // Direct tr under table (no thead/tbody)
-                        // This case is rare with html5ever as it auto-inserts tbody
-                        if headers.is_empty() {
-                            // First row is header
-                            self.extract_table_row_as_header(child, &mut headers, &mut alignments)?;
-                        } else {
-                            // Subsequent rows are data
-                            let mut row_cells = Vec::new();
-                            self.extract_table_row(child, &mut row_cells)?;
-                            rows.push(row_cells);
-                        }
-                    }
-                    _ => {
-                        // Ignore other elements
-                    }
-                }
-            }
-        }
-
-        // If no headers found, skip table conversion
-        if headers.is_empty() {
-            return Ok(());
-        }
-
-        // Ensure alignments match header count
-        while alignments.len() < headers.len() {
-            alignments.push(TableAlignment::Left);
-        }
-
-        // Generate GFM table
-        self.write_gfm_table(output, &headers, &alignments, &rows)?;
-
-        // Ensure blank line after table
-        if !output.ends_with("\n\n") {
-            output.push('\n');
-        }
-
-        Ok(())
-    }
-
-    /// Extract table header from thead element
-    fn extract_table_header(
-        &self,
-        thead: &Handle,
-        headers: &mut Vec<String>,
-        alignments: &mut Vec<TableAlignment>,
-    ) -> Result<(), ConversionError> {
-        // Find first tr in thead
-        for child in thead.children.borrow().iter() {
-            if let NodeData::Element { ref name, .. } = child.data
-                && name.local.as_ref() == "tr"
-            {
-                self.extract_table_row_as_header(child, headers, alignments)?;
-                break;
-            }
-        }
-        Ok(())
-    }
-
-    /// Extract table row as header (th elements)
-    fn extract_table_row_as_header(
-        &self,
-        tr: &Handle,
-        headers: &mut Vec<String>,
-        alignments: &mut Vec<TableAlignment>,
-    ) -> Result<(), ConversionError> {
-        for child in tr.children.borrow().iter() {
-            if let NodeData::Element {
-                ref name,
-                ref attrs,
-                ..
-            } = child.data
-            {
-                let tag = name.local.as_ref();
-                if tag == "th" || tag == "td" {
-                    // Extract cell content including inline formatting
-                    let mut cell_output = String::new();
-                    for cell_child in child.children.borrow().iter() {
-                        self.traverse_node(cell_child, &mut cell_output, 0)?;
-                    }
-                    // Normalize whitespace and trim
-                    let normalized = cell_output.trim().to_string();
-                    headers.push(normalized);
-
-                    // Extract alignment from attributes
-                    let attrs_borrowed = attrs.borrow();
-                    let alignment = self.extract_alignment(&attrs_borrowed);
-                    alignments.push(alignment);
-                }
-            }
-        }
-        Ok(())
-    }
-
-    /// Extract table rows from tbody element
-    fn extract_table_rows(
-        &self,
-        tbody: &Handle,
-        rows: &mut Vec<Vec<String>>,
-    ) -> Result<(), ConversionError> {
-        for child in tbody.children.borrow().iter() {
-            if let NodeData::Element { ref name, .. } = child.data
-                && name.local.as_ref() == "tr"
-            {
-                let mut row_cells = Vec::new();
-                self.extract_table_row(child, &mut row_cells)?;
-                rows.push(row_cells);
-            }
-        }
-        Ok(())
-    }
-
-    /// Extract cells from a table row
-    fn extract_table_row(
-        &self,
-        tr: &Handle,
-        cells: &mut Vec<String>,
-    ) -> Result<(), ConversionError> {
-        for child in tr.children.borrow().iter() {
-            if let NodeData::Element { ref name, .. } = child.data {
-                let tag = name.local.as_ref();
-                if tag == "td" || tag == "th" {
-                    // Extract cell content including inline formatting
-                    let mut cell_output = String::new();
-                    for cell_child in child.children.borrow().iter() {
-                        self.traverse_node(cell_child, &mut cell_output, 0)?;
-                    }
-                    // Normalize whitespace and trim
-                    let normalized = cell_output.trim().to_string();
-                    cells.push(normalized);
-                }
-            }
-        }
-        Ok(())
-    }
-
-    /// Extract alignment from element attributes
-    fn extract_alignment(&self, attrs: &Ref<Vec<html5ever::Attribute>>) -> TableAlignment {
-        // Check align attribute
-        for attr in attrs.iter() {
-            if attr.name.local.as_ref() == "align" {
-                let value = attr.value.to_string().to_lowercase();
-                return match value.as_str() {
-                    "left" => TableAlignment::Left,
-                    "center" => TableAlignment::Center,
-                    "right" => TableAlignment::Right,
-                    _ => TableAlignment::Left,
-                };
-            }
-        }
-
-        // Check style attribute for text-align
-        for attr in attrs.iter() {
-            if attr.name.local.as_ref() == "style" {
-                let style = attr.value.to_string().to_lowercase();
-                if style.contains("text-align") {
-                    if style.contains("center") {
-                        return TableAlignment::Center;
-                    } else if style.contains("right") {
-                        return TableAlignment::Right;
-                    } else if style.contains("left") {
-                        return TableAlignment::Left;
-                    }
-                }
-            }
-        }
-
-        TableAlignment::Left
-    }
-
-    /// Write GFM table to output
-    fn write_gfm_table(
-        &self,
-        output: &mut String,
-        headers: &[String],
-        alignments: &[TableAlignment],
-        rows: &[Vec<String>],
-    ) -> Result<(), ConversionError> {
-        // Write header row
-        output.push('|');
-        for header in headers {
-            output.push(' ');
-            output.push_str(header);
-            output.push_str(" |");
-        }
-        output.push('\n');
-
-        // Write separator row with alignment
-        output.push('|');
-        for alignment in alignments {
-            output.push(' ');
-            match alignment {
-                TableAlignment::Left => output.push_str("---"),
-                TableAlignment::Center => output.push_str(":---:"),
-                TableAlignment::Right => output.push_str("---:"),
-            }
-            output.push_str(" |");
-        }
-        output.push('\n');
-
-        // Write data rows
-        for row in rows {
-            output.push('|');
-            for (i, cell) in row.iter().enumerate() {
-                output.push(' ');
-                output.push_str(cell);
-                output.push_str(" |");
-
-                // If row has fewer cells than headers, pad with empty cells
-                if i >= headers.len() - 1 {
-                    break;
-                }
-            }
-            // Pad remaining cells if row is shorter than header
-            for _ in row.len()..headers.len() {
-                output.push_str("  |");
-            }
-            output.push('\n');
-        }
-
-        Ok(())
-    }
-
-    /// Extract code content from a node without any normalization
-    ///
-    /// This is critical for code blocks and inline code - we must preserve
-    /// the exact content including whitespace, line breaks, and indentation.
-    ///
-    /// # Arguments
-    ///
-    /// * `node` - The node to extract code from
-    /// * `output` - Mutable string buffer for code content
-    fn extract_code_content(
-        &self,
-        node: &Handle,
-        output: &mut String,
-    ) -> Result<(), ConversionError> {
-        match node.data {
-            NodeData::Text { ref contents } => {
-                // Add text content exactly as-is, NO normalization
-                output.push_str(&contents.borrow());
-            }
-            NodeData::Element { .. } => {
-                // Recursively extract from children
-                for child in node.children.borrow().iter() {
-                    self.extract_code_content(child, output)?;
-                }
-            }
-            _ => {
-                // Ignore other node types
-            }
-        }
-        Ok(())
-    }
-
-    /// Extract text content from a node and its descendants
-    ///
-    /// This helper function recursively extracts all text content from a node,
-    /// ignoring non-text elements. It's used to gather text for headings,
-    /// paragraphs, and other text-containing elements.
-    ///
-    /// # Arguments
-    ///
-    /// * `node` - The node to extract text from
-    /// * `output` - Mutable string buffer for extracted text
-    fn extract_text(&self, node: &Handle, output: &mut String) -> Result<(), ConversionError> {
-        match node.data {
-            NodeData::Text { ref contents } => {
-                output.push_str(&contents.borrow());
-            }
-            NodeData::Element { .. } => {
-                // Recursively extract text from children
-                for child in node.children.borrow().iter() {
-                    self.extract_text(child, output)?;
-                }
-            }
-            _ => {
-                // Ignore other node types
-            }
-        }
-        Ok(())
-    }
-
-    /// Normalize text content
-    ///
-    /// Applies text normalization rules to ensure consistent output:
-    /// - Collapses consecutive whitespace (spaces, tabs, newlines) to single spaces
-    /// - Trims leading and trailing whitespace
-    /// - Preserves intentional line breaks (future enhancement)
-    ///
-    /// # Arguments
-    ///
-    /// * `text` - Raw text content to normalize
-    ///
-    /// # Returns
-    ///
-    /// Normalized text string
-    ///
-    /// # Examples
-    ///
-    /// ```text
-    /// "  multiple   spaces  " -> "multiple spaces"
-    /// "line\nbreak" -> "line break"
-    /// "  \t  tabs  \t  " -> "tabs"
-    /// ```
-    fn normalize_text(&self, text: &str) -> String {
-        // Split on whitespace and filter empty strings
-        let words: Vec<&str> = text.split_whitespace().collect();
-
-        // Join with single spaces
-        words.join(" ")
-    }
-
-    /// Normalize final output for deterministic Markdown generation
-    ///
-    /// Applies comprehensive normalization to ensure deterministic output for stable ETags:
-    ///
-    /// **Normalization Rules:**
-    /// 1. **Line Endings**: Enforce LF (`\n`) only, never CRLF (`\r\n`)
-    /// 2. **Blank Lines**: Collapse consecutive blank lines to single blank line
-    /// 3. **Trailing Whitespace**: Remove trailing whitespace from all lines
-    /// 4. **Final Newline**: Ensure exactly one newline at end of file
-    /// 5. **Whitespace Normalization**: Collapse consecutive spaces to single space
-    /// 6. **Markdown Escaping**: Apply consistent escaping rules for special characters
-    ///
-    /// These rules ensure that converting identical HTML twice produces identical Markdown,
-    /// which is critical for stable ETag generation and predictable caching behavior.
-    ///
-    /// # Arguments
-    ///
-    /// * `output` - Raw Markdown output
-    ///
-    /// # Returns
-    ///
-    /// Normalized Markdown string with deterministic formatting
-    ///
-    /// # Examples
-    ///
-    /// ```
-    /// // Input with CRLF and multiple blank lines
-    /// let input = "Line 1\r\n\r\n\r\nLine 2  \n";
-    /// // Output with LF and single blank line
-    /// let output = "Line 1\n\nLine 2\n";
-    /// ```
-    fn normalize_output(&self, output: String) -> String {
-        // Step 1: Normalize line endings (CRLF -> LF)
-        let output = output.replace("\r\n", "\n");
-
-        // Step 2: Normalize whitespace within lines (collapse consecutive spaces)
-        // This is done line-by-line to preserve intentional spacing in code blocks
-        let mut result = String::with_capacity(output.len());
-        let mut prev_blank = false;
-        let mut in_code_block = false;
-
-        for line in output.lines() {
-            // Detect code block boundaries (fenced code blocks start with ```)
-            if line.trim_start().starts_with("```") {
-                in_code_block = !in_code_block;
-            }
-
-            // Step 3: Remove trailing whitespace from all lines
-            let trimmed = line.trim_end();
-
-            if trimmed.is_empty() {
-                // Step 4: Collapse consecutive blank lines to single blank line
-                if !prev_blank {
-                    result.push('\n');
-                    prev_blank = true;
-                }
-            } else {
-                // Step 5: Normalize whitespace (collapse consecutive spaces)
-                // Skip normalization inside code blocks to preserve formatting
-                if in_code_block {
-                    result.push_str(trimmed);
-                } else {
-                    // Collapse consecutive spaces to single space
-                    let normalized = self.normalize_line_whitespace(trimmed);
-                    result.push_str(&normalized);
-                }
-                result.push('\n');
-                prev_blank = false;
-            }
-        }
-
-        // Step 6: Ensure single trailing newline
-        if !result.ends_with('\n') {
-            result.push('\n');
-        } else if result.ends_with("\n\n") {
-            // Remove extra trailing newlines
-            while result.ends_with("\n\n") {
-                result.pop();
-            }
-        }
-
-        result
-    }
-
-    /// Normalize whitespace within a single line
-    ///
-    /// Collapses consecutive spaces to a single space while preserving
-    /// intentional spacing in Markdown syntax (e.g., list indentation, inline code).
-    ///
-    /// # Arguments
-    ///
-    /// * `line` - A single line of text
-    ///
-    /// # Returns
-    ///
-    /// Line with normalized whitespace
-    fn normalize_line_whitespace(&self, line: &str) -> String {
-        let mut result = String::with_capacity(line.len());
-        let mut prev_space = false;
-        let mut at_start = true;
-        let mut in_inline_code = false;
-
-        for ch in line.chars() {
-            if ch == '`' {
-                // Toggle inline code state
-                in_inline_code = !in_inline_code;
-                result.push(ch);
-                prev_space = false;
-                at_start = false;
-            } else if ch == ' ' {
-                if in_inline_code {
-                    // Preserve all spaces inside inline code
-                    result.push(ch);
-                } else if at_start {
-                    // Preserve leading spaces (for list indentation)
-                    result.push(ch);
-                } else if !prev_space {
-                    // First space in a sequence
-                    result.push(ch);
-                    prev_space = true;
-                }
-                // Skip consecutive spaces (unless at start or in code)
-            } else {
-                result.push(ch);
-                prev_space = false;
-                at_start = false;
-            }
-        }
-
-        result
     }
 }
 
@@ -4268,378 +2535,5 @@ mod tests {
             result.contains("\u{201D}"),
             "Right double quote should be decoded"
         );
-    }
-
-    // ============================================================================
-    // Table Conversion Tests (GFM)
-    // ============================================================================
-
-    /// Test basic table conversion with GFM flavor
-    /// Validates: FR-11.2
-    #[test]
-    fn test_table_basic_gfm() {
-        let html = b"<table><thead><tr><th>Header 1</th><th>Header 2</th></tr></thead><tbody><tr><td>Cell 1</td><td>Cell 2</td></tr></tbody></table>";
-        let dom = parse_html(html).expect("Parse failed");
-
-        let options = ConversionOptions {
-            flavor: MarkdownFlavor::GitHubFlavoredMarkdown,
-            ..Default::default()
-        };
-        let converter = MarkdownConverter::with_options(options);
-        let result = converter.convert(&dom).expect("Conversion failed");
-
-        // Should contain GFM table format
-        assert!(result.contains("| Header 1 | Header 2 |"));
-        assert!(result.contains("| --- | --- |"));
-        assert!(result.contains("| Cell 1 | Cell 2 |"));
-    }
-
-    /// Test that tables are NOT converted with CommonMark flavor
-    /// Validates: FR-11.2
-    #[test]
-    fn test_table_not_converted_commonmark() {
-        let html = b"<table><thead><tr><th>Header</th></tr></thead><tbody><tr><td>Cell</td></tr></tbody></table>";
-        let dom = parse_html(html).expect("Parse failed");
-
-        let options = ConversionOptions {
-            flavor: MarkdownFlavor::CommonMark,
-            ..Default::default()
-        };
-        let converter = MarkdownConverter::with_options(options);
-        let result = converter.convert(&dom).expect("Conversion failed");
-
-        // Should NOT contain GFM table format
-        assert!(!result.contains("|"));
-        // Should contain the text content
-        assert!(result.contains("Header"));
-        assert!(result.contains("Cell"));
-    }
-
-    /// Test table with left alignment (default)
-    /// Validates: FR-11.2
-    #[test]
-    fn test_table_left_alignment() {
-        let html = b"<table><thead><tr><th align=\"left\">Left</th></tr></thead><tbody><tr><td>Data</td></tr></tbody></table>";
-        let dom = parse_html(html).expect("Parse failed");
-
-        let options = ConversionOptions {
-            flavor: MarkdownFlavor::GitHubFlavoredMarkdown,
-            ..Default::default()
-        };
-        let converter = MarkdownConverter::with_options(options);
-        let result = converter.convert(&dom).expect("Conversion failed");
-
-        // Left alignment uses default separator
-        assert!(result.contains("| --- |"));
-        assert!(result.contains("| Left |"));
-    }
-
-    /// Test table with center alignment
-    /// Validates: FR-11.2
-    #[test]
-    fn test_table_center_alignment() {
-        let html = b"<table><thead><tr><th align=\"center\">Center</th></tr></thead><tbody><tr><td>Data</td></tr></tbody></table>";
-        let dom = parse_html(html).expect("Parse failed");
-
-        let options = ConversionOptions {
-            flavor: MarkdownFlavor::GitHubFlavoredMarkdown,
-            ..Default::default()
-        };
-        let converter = MarkdownConverter::with_options(options);
-        let result = converter.convert(&dom).expect("Conversion failed");
-
-        // Center alignment uses :---:
-        assert!(result.contains("| :---: |"));
-        assert!(result.contains("| Center |"));
-    }
-
-    /// Test table with right alignment
-    /// Validates: FR-11.2
-    #[test]
-    fn test_table_right_alignment() {
-        let html = b"<table><thead><tr><th align=\"right\">Right</th></tr></thead><tbody><tr><td>Data</td></tr></tbody></table>";
-        let dom = parse_html(html).expect("Parse failed");
-
-        let options = ConversionOptions {
-            flavor: MarkdownFlavor::GitHubFlavoredMarkdown,
-            ..Default::default()
-        };
-        let converter = MarkdownConverter::with_options(options);
-        let result = converter.convert(&dom).expect("Conversion failed");
-
-        // Right alignment uses ---:
-        assert!(result.contains("| ---: |"));
-        assert!(result.contains("| Right |"));
-    }
-
-    /// Test table with mixed alignments
-    /// Validates: FR-11.2
-    #[test]
-    fn test_table_mixed_alignments() {
-        let html = b"<table><thead><tr><th align=\"left\">Left</th><th align=\"center\">Center</th><th align=\"right\">Right</th></tr></thead><tbody><tr><td>A</td><td>B</td><td>C</td></tr></tbody></table>";
-        let dom = parse_html(html).expect("Parse failed");
-
-        let options = ConversionOptions {
-            flavor: MarkdownFlavor::GitHubFlavoredMarkdown,
-            ..Default::default()
-        };
-        let converter = MarkdownConverter::with_options(options);
-        let result = converter.convert(&dom).expect("Conversion failed");
-
-        // Should have mixed alignment separators
-        assert!(result.contains("| --- | :---: | ---: |"));
-        assert!(result.contains("| Left | Center | Right |"));
-        assert!(result.contains("| A | B | C |"));
-    }
-
-    /// Test table with style-based alignment
-    /// Validates: FR-11.2
-    #[test]
-    fn test_table_style_alignment() {
-        let html = b"<table><thead><tr><th style=\"text-align: center\">Styled</th></tr></thead><tbody><tr><td>Data</td></tr></tbody></table>";
-        let dom = parse_html(html).expect("Parse failed");
-
-        let options = ConversionOptions {
-            flavor: MarkdownFlavor::GitHubFlavoredMarkdown,
-            ..Default::default()
-        };
-        let converter = MarkdownConverter::with_options(options);
-        let result = converter.convert(&dom).expect("Conversion failed");
-
-        // Should detect alignment from style attribute
-        assert!(result.contains("| :---: |"));
-    }
-
-    /// Test table without thead (direct tr under table)
-    /// Validates: FR-11.2
-    #[test]
-    fn test_table_without_thead() {
-        let html = b"<table><tr><th>Header 1</th><th>Header 2</th></tr><tr><td>Cell 1</td><td>Cell 2</td></tr></table>";
-        let dom = parse_html(html).expect("Parse failed");
-
-        let options = ConversionOptions {
-            flavor: MarkdownFlavor::GitHubFlavoredMarkdown,
-            ..Default::default()
-        };
-        let converter = MarkdownConverter::with_options(options);
-        let result = converter.convert(&dom).expect("Conversion failed");
-
-        // Should still convert properly
-        assert!(result.contains("| Header 1 | Header 2 |"));
-        assert!(result.contains("| --- | --- |"));
-        assert!(result.contains("| Cell 1 | Cell 2 |"));
-    }
-
-    /// Test table with multiple rows
-    /// Validates: FR-11.2
-    #[test]
-    fn test_table_multiple_rows() {
-        let html = b"<table><thead><tr><th>Name</th><th>Age</th></tr></thead><tbody><tr><td>Alice</td><td>30</td></tr><tr><td>Bob</td><td>25</td></tr><tr><td>Charlie</td><td>35</td></tr></tbody></table>";
-        let dom = parse_html(html).expect("Parse failed");
-
-        let options = ConversionOptions {
-            flavor: MarkdownFlavor::GitHubFlavoredMarkdown,
-            ..Default::default()
-        };
-        let converter = MarkdownConverter::with_options(options);
-        let result = converter.convert(&dom).expect("Conversion failed");
-
-        // Should have all rows
-        assert!(result.contains("| Name | Age |"));
-        assert!(result.contains("| Alice | 30 |"));
-        assert!(result.contains("| Bob | 25 |"));
-        assert!(result.contains("| Charlie | 35 |"));
-    }
-
-    /// Test table with empty cells
-    /// Validates: FR-11.2
-    #[test]
-    fn test_table_empty_cells() {
-        let html = b"<table><thead><tr><th>Col1</th><th>Col2</th></tr></thead><tbody><tr><td>Data</td><td></td></tr><tr><td></td><td>Data</td></tr></tbody></table>";
-        let dom = parse_html(html).expect("Parse failed");
-
-        let options = ConversionOptions {
-            flavor: MarkdownFlavor::GitHubFlavoredMarkdown,
-            ..Default::default()
-        };
-        let converter = MarkdownConverter::with_options(options);
-        let result = converter.convert(&dom).expect("Conversion failed");
-
-        // Should handle empty cells
-        assert!(result.contains("| Col1 | Col2 |"));
-        assert!(result.contains("| Data | |"));
-        assert!(result.contains("| | Data |"));
-    }
-
-    /// Test table with uneven rows (fewer cells than headers)
-    /// Validates: FR-11.2
-    #[test]
-    fn test_table_uneven_rows() {
-        let html = b"<table><thead><tr><th>A</th><th>B</th><th>C</th></tr></thead><tbody><tr><td>1</td><td>2</td></tr><tr><td>3</td></tr></tbody></table>";
-        let dom = parse_html(html).expect("Parse failed");
-
-        let options = ConversionOptions {
-            flavor: MarkdownFlavor::GitHubFlavoredMarkdown,
-            ..Default::default()
-        };
-        let converter = MarkdownConverter::with_options(options);
-        let result = converter.convert(&dom).expect("Conversion failed");
-
-        // Should pad missing cells
-        assert!(result.contains("| A | B | C |"));
-        assert!(result.contains("| 1 | 2 | |"));
-        assert!(result.contains("| 3 | | |"));
-    }
-
-    /// Test table with text formatting in cells
-    /// Validates: FR-11.2
-    #[test]
-    fn test_table_with_formatting() {
-        let html = b"<table><thead><tr><th>Name</th><th>Status</th></tr></thead><tbody><tr><td><strong>Bold</strong></td><td><em>Italic</em></td></tr></tbody></table>";
-        let dom = parse_html(html).expect("Parse failed");
-
-        let options = ConversionOptions {
-            flavor: MarkdownFlavor::GitHubFlavoredMarkdown,
-            ..Default::default()
-        };
-        let converter = MarkdownConverter::with_options(options);
-        let result = converter.convert(&dom).expect("Conversion failed");
-
-        // Should preserve formatting in cells
-        assert!(result.contains("| Name | Status |"));
-        assert!(result.contains("| **Bold** | *Italic* |"));
-    }
-
-    /// Test table with links in cells
-    /// Validates: FR-11.2
-    #[test]
-    fn test_table_with_links() {
-        let html = b"<table><thead><tr><th>Site</th></tr></thead><tbody><tr><td><a href=\"https://example.com\">Example</a></td></tr></tbody></table>";
-        let dom = parse_html(html).expect("Parse failed");
-
-        let options = ConversionOptions {
-            flavor: MarkdownFlavor::GitHubFlavoredMarkdown,
-            ..Default::default()
-        };
-        let converter = MarkdownConverter::with_options(options);
-        let result = converter.convert(&dom).expect("Conversion failed");
-
-        // Should preserve links in cells
-        assert!(result.contains("| Site |"));
-        assert!(result.contains("| [Example](https://example.com) |"));
-    }
-
-    /// Test table with code in cells
-    /// Validates: FR-11.2
-    #[test]
-    fn test_table_with_code() {
-        let html = b"<table><thead><tr><th>Function</th></tr></thead><tbody><tr><td><code>print()</code></td></tr></tbody></table>";
-        let dom = parse_html(html).expect("Parse failed");
-
-        let options = ConversionOptions {
-            flavor: MarkdownFlavor::GitHubFlavoredMarkdown,
-            ..Default::default()
-        };
-        let converter = MarkdownConverter::with_options(options);
-        let result = converter.convert(&dom).expect("Conversion failed");
-
-        // Should preserve inline code in cells
-        assert!(result.contains("| Function |"));
-        assert!(result.contains("| `print()` |"));
-    }
-
-    /// Test table blank line separation
-    /// Validates: FR-11.2
-    #[test]
-    fn test_table_blank_line_separation() {
-        let html = b"<p>Before table</p><table><thead><tr><th>Header</th></tr></thead><tbody><tr><td>Data</td></tr></tbody></table><p>After table</p>";
-        let dom = parse_html(html).expect("Parse failed");
-
-        let options = ConversionOptions {
-            flavor: MarkdownFlavor::GitHubFlavoredMarkdown,
-            ..Default::default()
-        };
-        let converter = MarkdownConverter::with_options(options);
-        let result = converter.convert(&dom).expect("Conversion failed");
-
-        // Should have blank lines around table
-        assert!(result.contains("Before table\n\n|"));
-        assert!(result.contains("|\n\nAfter table"));
-    }
-
-    /// Test table with no tbody (only thead)
-    /// Validates: FR-11.2
-    #[test]
-    fn test_table_thead_only() {
-        let html = b"<table><thead><tr><th>Header</th></tr></thead></table>";
-        let dom = parse_html(html).expect("Parse failed");
-
-        let options = ConversionOptions {
-            flavor: MarkdownFlavor::GitHubFlavoredMarkdown,
-            ..Default::default()
-        };
-        let converter = MarkdownConverter::with_options(options);
-        let result = converter.convert(&dom).expect("Conversion failed");
-
-        // Should still generate table with header and separator
-        assert!(result.contains("| Header |"));
-        assert!(result.contains("| --- |"));
-    }
-
-    /// Test table with td in header row (some HTML uses td instead of th)
-    /// Validates: FR-11.2
-    #[test]
-    fn test_table_td_as_header() {
-        let html = b"<table><tr><td>Header 1</td><td>Header 2</td></tr><tr><td>Cell 1</td><td>Cell 2</td></tr></table>";
-        let dom = parse_html(html).expect("Parse failed");
-
-        let options = ConversionOptions {
-            flavor: MarkdownFlavor::GitHubFlavoredMarkdown,
-            ..Default::default()
-        };
-        let converter = MarkdownConverter::with_options(options);
-        let result = converter.convert(&dom).expect("Conversion failed");
-
-        // First row should be treated as header
-        assert!(result.contains("| Header 1 | Header 2 |"));
-        assert!(result.contains("| --- | --- |"));
-        assert!(result.contains("| Cell 1 | Cell 2 |"));
-    }
-
-    /// Test empty table (no headers)
-    /// Validates: FR-11.2
-    #[test]
-    fn test_table_empty() {
-        let html = b"<table></table>";
-        let dom = parse_html(html).expect("Parse failed");
-
-        let options = ConversionOptions {
-            flavor: MarkdownFlavor::GitHubFlavoredMarkdown,
-            ..Default::default()
-        };
-        let converter = MarkdownConverter::with_options(options);
-        let result = converter.convert(&dom).expect("Conversion failed");
-
-        // Empty table should not produce output
-        assert!(!result.contains("|"));
-    }
-
-    /// Test table with whitespace in cells
-    /// Validates: FR-11.2
-    #[test]
-    fn test_table_whitespace_normalization() {
-        let html = b"<table><thead><tr><th>  Header  </th></tr></thead><tbody><tr><td>  Data  with   spaces  </td></tr></tbody></table>";
-        let dom = parse_html(html).expect("Parse failed");
-
-        let options = ConversionOptions {
-            flavor: MarkdownFlavor::GitHubFlavoredMarkdown,
-            ..Default::default()
-        };
-        let converter = MarkdownConverter::with_options(options);
-        let result = converter.convert(&dom).expect("Conversion failed");
-
-        // Whitespace should be normalized
-        assert!(result.contains("| Header |"));
-        assert!(result.contains("| Data with spaces |"));
     }
 }
