+++ components/rust-converter/src/ffi/exports.rs
@@ -1,0 +1,101 @@
+use std::panic;
+use std::ptr;
+
+use crate::error::ConversionError;
+
+use super::abi::{ERROR_INTERNAL, MarkdownConverterHandle, MarkdownOptions, MarkdownResult};
+use super::convert::convert_inner;
+use super::memory::{free_buffer, reset_result, set_error_result, set_success_result};
+use super::options::{required_bytes, required_ref};
+
+#[unsafe(no_mangle)]
+pub extern "C" fn markdown_converter_new() -> *mut MarkdownConverterHandle {
+    let result = panic::catch_unwind(|| Box::into_raw(Box::new(MarkdownConverterHandle::new())));
+    match result {
+        Ok(ptr) => ptr,
+        Err(_) => ptr::null_mut(),
+    }
+}
+
+/// Convert an HTML byte buffer into Markdown using the provided converter handle.
+///
+/// # Safety
+///
+/// The caller must ensure that:
+/// - `handle` points to a live converter created by `markdown_converter_new()`
+/// - `options` points to a valid `MarkdownOptions`
+/// - `result` points to writable storage for a `MarkdownResult`
+/// - `html` either points to `html_len` readable bytes or is NULL when `html_len == 0`
+/// - `result` is not concurrently mutated while this function is executing
+#[unsafe(no_mangle)]
+pub unsafe extern "C" fn markdown_convert(
+    handle: *mut MarkdownConverterHandle,
+    html: *const u8,
+    html_len: usize,
+    options: *const MarkdownOptions,
+    result: *mut MarkdownResult,
+) {
+    if result.is_null() {
+        return;
+    }
+
+    // SAFETY: `result` was validated as non-NULL above.
+    let result_ref = unsafe { &mut *result };
+    reset_result(result_ref);
+
+    let panic_result = panic::catch_unwind(|| -> Result<_, ConversionError> {
+        let handle_ref = required_ref(handle.cast_const(), "Converter handle")?;
+        let options_ref = required_ref(options, "Options")?;
+        let html_slice = required_bytes(html, html_len, "HTML")?;
+        convert_inner(handle_ref, html_slice, options_ref)
+    });
+
+    match panic_result {
+        Ok(Ok(output)) => set_success_result(result_ref, output),
+        Ok(Err(err)) => set_error_result(result_ref, err.code(), err.to_string()),
+        Err(_) => set_error_result(
+            result_ref,
+            ERROR_INTERNAL,
+            "Internal panic during conversion".to_string(),
+        ),
+    }
+}
+
+/// Release buffers owned by a `MarkdownResult` and reset it to the empty state.
+///
+/// # Safety
+///
+/// The caller must ensure that `result` either is NULL or points to a valid
+/// `MarkdownResult` previously initialized by `markdown_convert()`. Passing the
+/// same result twice is allowed because the function resets pointers to NULL.
+#[unsafe(no_mangle)]
+pub unsafe extern "C" fn markdown_result_free(result: *mut MarkdownResult) {
+    if result.is_null() {
+        return;
+    }
+
+    // SAFETY: `result` was validated as non-NULL above.
+    let result_ref = unsafe { &mut *result };
+    free_buffer(&mut result_ref.markdown, &mut result_ref.markdown_len);
+    free_buffer(&mut result_ref.etag, &mut result_ref.etag_len);
+    free_buffer(&mut result_ref.error_message, &mut result_ref.error_len);
+    result_ref.token_estimate = 0;
+    result_ref.error_code = 0;
+}
+
+/// Destroy a converter handle previously returned by `markdown_converter_new()`.
+///
+/// # Safety
+///
+/// The caller must ensure that `handle` either is NULL or was returned by
+/// `markdown_converter_new()` and has not already been freed. The handle must
+/// not be used after this call returns.
+#[unsafe(no_mangle)]
+pub unsafe extern "C" fn markdown_converter_free(handle: *mut MarkdownConverterHandle) {
+    if handle.is_null() {
+        return;
+    }
+
+    // SAFETY: `handle` was validated as non-NULL above and originated from `Box::into_raw`.
+    unsafe { drop(Box::from_raw(handle)) };
+}
